requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
aiohttp-jinja2==1.4.2
    # via -r requirements.in
aiohttp==3.7.3
    # via
    #   -r requirements.in
    #   aiohttp-jinja2
    #   discord.py
appdirs==1.4.4
    # via black
argh==0.26.2
    # via watchdog
astroid==2.4.2
    # via pylint
async-timeout==3.0.1
    # via aiohttp
attrs==20.3.0
    # via
    #   aiohttp
    #   pytest
black==20.8b1
    # via -r requirements.in
bottle==0.12.19
    # via platformio
certifi==2020.12.5
    # via
    #   requests
    #   sentry-sdk
chardet==3.0.4
    # via
    #   aiohttp
    #   requests
click==7.1.2
    # via
    #   black
    #   platformio
colorama==0.4.4
    # via platformio
deprecated==1.2.11
    # via pygithub
discord.py==1.6.0
    # via -r requirements.in
gitdb==4.0.5
    # via gitpython
gitpython==3.1.12
    # via -r requirements.in
humanize==3.2.0
    # via -r requirements.in
idna==2.10
    # via
    #   requests
    #   yarl
iniconfig==1.1.1
    # via pytest
isort==5.7.0
    # via pylint
jinja2==2.11.2
    # via aiohttp-jinja2
lazy-object-proxy==1.4.3
    # via astroid
loguru==0.5.3
    # via -r requirements.in
markupsafe==1.1.1
    # via jinja2
marshmallow==3.10.0
    # via platformio
mccabe==0.6.1
    # via pylint
multidict==5.1.0
    # via
    #   aiohttp
    #   yarl
mypy-extensions==0.4.3
    # via black
packaging==20.8
    # via pytest
pathspec==0.8.1
    # via black
platformio==5.0.4
    # via -r requirements.in
pluggy==0.13.1
    # via pytest
py==1.10.0
    # via pytest
pydantic==1.7.3
    # via -r requirements.in
pyelftools==0.27
    # via platformio
pygithub==1.54.1
    # via -r requirements.in
pyjwt==1.7.1
    # via pygithub
pylint==2.6.0
    # via -r requirements.in
pyparsing==2.4.7
    # via packaging
pyserial==3.5
    # via platformio
pytest==6.2.1
    # via -r requirements.in
python-dotenv==0.15.0
    # via -r requirements.in
pyyaml==5.4
    # via watchdog
regex==2020.11.13
    # via black
requests==2.25.1
    # via
    #   platformio
    #   pygithub
semantic-version==2.8.5
    # via platformio
sentry-sdk==0.19.5
    # via -r requirements.in
shortuuid==1.0.1
    # via -r requirements.in
six==1.15.0
    # via astroid
smmap==3.0.4
    # via gitdb
tabulate==0.8.7
    # via platformio
toml==0.10.2
    # via
    #   black
    #   pylint
    #   pytest
typed-ast==1.4.2
    # via black
typing-extensions==3.7.4.3
    # via
    #   aiohttp
    #   aiohttp-jinja2
    #   black
urllib3==1.26.2
    # via
    #   requests
    #   sentry-sdk
watchdog[watchmedo]==1.0.2
    # via -r requirements.in
wrapt==1.12.1
    # via
    #   astroid
    #   deprecated
yarl==1.6.3
    # via aiohttp

# The following packages are considered to be unsafe in a requirements file:
# setuptools
